hydra: move lib/pmi_v2_common to internal
The routines in lib/pmi_v2_common is only a small portion of the kvs
queuing logic. Move them into static routines in mpiexec/pmiserv_kvs.c
and proxy/pmip_pmi.c allows the code logic be contained together. It
also allows to tailor data structure and interface more specific to mpiexec
and proxy respectively.
