+++ src/pm/hydra/lib/Makefile.mk
@@ -6,8 +6,7 @@
 AM_CPPFLAGS += -I$(top_srcdir)/lib
 
 libhydra_la_SOURCES += \
-    lib/pmiserv_common.c \
-    lib/pmi_v2_common.c
+    lib/pmiserv_common.c
 
 include lib/utils/Makefile.mk
 include lib/tools/Makefile.mk
