+++ src/pm/hydra/mpiexec/pmiserv_kvs.c
@@ -6,11 +6,23 @@
 #include "hydra_server.h"
 #include "pmiserv_pmi.h"
 #include "pmiserv_utils.h"
-#include "pmi_v2_common.h"
+
+struct HYD_pmcd_pmi_v2_reqs {
+    int fd;
+    int pid;
+    int pgid;
+    struct PMIU_cmd *pmi;
+    const char *key;
+
+    struct HYD_pmcd_pmi_v2_reqs *prev;
+    struct HYD_pmcd_pmi_v2_reqs *next;
+};
 
 static struct HYD_pmcd_pmi_v2_reqs *pending_reqs = NULL;
 
 static bool check_epoch_reached(struct HYD_pg *pg, int fd, int pid);
+static HYD_status HYD_pmcd_pmi_v2_queue_req(int fd, int pid, int pgid,
+                                            struct PMIU_cmd *pmi, const char *key);
 static HYD_status check_pending_reqs(const char *key);
 
 static const bool is_static = true;
@@ -60,7 +72,7 @@ HYD_status HYD_pmiserv_kvs_get(int fd, int pid, int pgid, struct PMIU_cmd *pmi, 
          * the "get". A "put" (from another proxy) will poke the progress.
          */
         if (!check_epoch_reached(proxy->pg, fd, pid)) {
-            status = HYD_pmcd_pmi_v2_queue_req(fd, pid, pgid, pmi, key, &pending_reqs);
+            status = HYD_pmcd_pmi_v2_queue_req(fd, pid, pgid, pmi, key);
             HYDU_ERR_POP(status, "unable to queue request\n");
             goto fn_exit;
         }
@@ -282,6 +294,38 @@ static bool check_epoch_reached(struct HYD_pg *pg, int fd, int pid)
     }
 
     return true;
+}
+
+static HYD_status HYD_pmcd_pmi_v2_queue_req(int fd, int pid, int pgid,
+                                            struct PMIU_cmd *pmi, const char *key)
+{
+    struct HYD_pmcd_pmi_v2_reqs *req, *tmp;
+    HYD_status status = HYD_SUCCESS;
+
+    HYDU_MALLOC_OR_JUMP(req, struct HYD_pmcd_pmi_v2_reqs *, sizeof(struct HYD_pmcd_pmi_v2_reqs),
+                        status);
+    req->fd = fd;
+    req->pid = pid;
+    req->pgid = pgid;
+    req->prev = NULL;
+    req->next = NULL;
+
+    req->pmi = PMIU_cmd_dup(pmi);
+    req->key = MPL_strdup(key);
+
+    if (pending_reqs == NULL)
+        pending_reqs = req;
+    else {
+        for (tmp = pending_reqs; tmp->next; tmp = tmp->next);
+        tmp->next = req;
+        req->prev = tmp;
+    }
+
+  fn_exit:
+    return status;
+
+  fn_fail:
+    goto fn_exit;
 }
 
 /* Process the pending get with matching key. If the key is NULL,
