+++ src/pm/hydra/proxy/pmip_pmi.c
@@ -9,7 +9,6 @@
 #include "demux.h"
 #include "topo.h"
 #include "uthash.h"
-#include "pmi_v2_common.h"
 
 #define debug(...)                              \
     {                                           \
@@ -57,8 +56,48 @@ static void internal_finalize(void)
     MPL_free(cache_get);
 }
 
+/* info_getnodeattr will wait for info_putnodeattr */
+struct HYD_pmcd_pmi_v2_reqs {
+    int fd;
+    struct PMIU_cmd *pmi;
+    const char *key;
+
+    struct HYD_pmcd_pmi_v2_reqs *prev;
+    struct HYD_pmcd_pmi_v2_reqs *next;
+};
+
 static struct HYD_pmcd_pmi_v2_reqs *pending_reqs = NULL;
 
+static HYD_status HYD_pmcd_pmi_v2_queue_req(int fd, struct PMIU_cmd *pmi, const char *key)
+{
+    struct HYD_pmcd_pmi_v2_reqs *req, *tmp;
+    HYD_status status = HYD_SUCCESS;
+
+    HYDU_MALLOC_OR_JUMP(req, struct HYD_pmcd_pmi_v2_reqs *, sizeof(struct HYD_pmcd_pmi_v2_reqs),
+                        status);
+    req->fd = fd;
+    req->prev = NULL;
+    req->next = NULL;
+
+    req->pmi = PMIU_cmd_dup(pmi);
+    req->key = MPL_strdup(key);
+
+    if (pending_reqs == NULL)
+        pending_reqs = req;
+    else {
+        for (tmp = pending_reqs; tmp->next; tmp = tmp->next);
+        tmp->next = req;
+        req->prev = tmp;
+    }
+
+  fn_exit:
+    return status;
+
+  fn_fail:
+    goto fn_exit;
+}
+
+/* check pending kvs get when we get fn_info_putnodeattr */
 static HYD_status poke_progress(const char *key)
 {
     struct HYD_pmcd_pmi_v2_reqs *req, *list_head = NULL, *list_tail = NULL;
@@ -678,7 +717,7 @@ HYD_status fn_info_getnodeattr(int fd, struct PMIU_cmd *pmi)
     }
 
     if (!found && wait) {
-        status = HYD_pmcd_pmi_v2_queue_req(fd, -1, -1, pmi, key, &pending_reqs);
+        status = HYD_pmcd_pmi_v2_queue_req(fd, pmi, key);
         HYDU_ERR_POP(status, "unable to queue request\n");
         goto fn_exit;
     }
